* Finalizada a documentacao de LadderObject.h * Alterado ponteiro global DrawChars para ponteiro na classe LadderDiagram, de forma que a classe nao dependa de declaracoes externas a ela * Corrigidos warnings devido a conversao int -> float na funcao DrawLine de EngineRendererD2D * Atualizado o exemplo abs_rand.ld para o novo formato * Alterado o exemplo math.ld para conter instrucao de resto de divisao + Adicionados comandos PageUp / PageDown para movimentar a tela do ladder. corrigido #1224 * Corrigida funcao que altera o tipo de I/O que em alguns casos nao atualizava a barra de titulos do ladder * Alterada janela de configuracao do nome da bobina para exibir a lista de I/Os de todos os tipos suportados quando for do tipo reservado * Alterada funcao que atualiza o tipo de I/O da bobina para nao dar mensagem de erro quando o tipo do I/O for reservado * Alterada funcao que cria linha no diagrama para retornar o estado da operacao: sucesso ou falha * Movidas variaveis que armazenam o offset da tela de draw_outputdev.cpp para maincontrols.cpp

+++ EngineRenderD2D.cpp
@@ -1028,8 +1028,8 @@ HRESULT EngineRenderD2D::DrawLine(POINT start, POINT end, unsigned int brush, un
 	if(pRT != NULL && pBrush != nullptr) {
 		D2D1_POINT_2F center;
 		if(angle) {
-			center.x = start.x + (end.x - start.x)/2;
-			center.y = start.y + (end.y - start.y)/2;
+			center.x = (FLOAT)(start.x + (FLOAT)(end.x - start.x)/2);
+			center.y = (FLOAT)(start.y + (FLOAT)(end.y - start.y)/2);
 			pRT->SetTransform(D2D1::Matrix3x2F::Rotation((float)angle, center));
 		}
